refactor(pacbio): simplify sample sheet behaviour retrieval and remove deprecated classes
remove Json APi processor for load_included
