+++ app/exchanges/run_csv/pacbio_sample_sheet.rb
@@ -1,71 +1,129 @@
 # frozen_string_literal: true
 
+# A class spefically for SMRT-Link v13 Sample Sheets, based on the PacbioSampleSheet class
+# See https://www.pacb.com/wp-content/uploads/SMRT-Link-User-Guide-v13.1.pdf (page 31) for details.
+
 module RunCsv
   # RunCsv::PacbioSampleSheet
-  # Used to generate sample sheets specific to the PacBio pipeline for v12 and above
-  class PacbioSampleSheet < DataStructureBuilder
-    # return a list of column names ie headers
-    # eg ['System Name', 'Run Name']
-    def csv_headers
-      configuration.column_order
+  # Used to generate sample sheets specific to the PacBio pipeline (introduced in SMRT Link v13)
+  class PacbioSampleSheet < PacbioSampleSheetLegacy
+    # Generate a hash of settings for the run
+    def run_settings
+      run = object # Pacbio::Run
+
+      plate_data = run.plates.first(2).each_with_index.with_object({}) do |(plate, index), hash|
+        hash["Plate #{index + 1}"] = plate&.sequencing_kit_box_barcode
+      end
+
+      {
+        'Instrument Type' => run.system_name,
+        'Run Name' =>	run.name,
+        'Run Comments' =>	run.comments
+      }.merge(plate_data).merge(
+        { 'CSV Version' => 1 }
+      )
     end
 
-    # Returns true if the object can be iterated using `each`,
-    # false otherwise
-    def eachable?(object)
-      object.respond_to?(:each)
+    # Generate a list of plate-well identifiers.
+    # Eg. ['1_A01', '1_A02', '2_A01', ...]
+    def plate_well_names
+      wells = object.plates.flat_map(&:wells)
+      wells.map do |well|
+        "#{well.plate.plate_number}_#{well.position_leading_zero}"
+      end
     end
 
-    # Flattens a nested hash of arrays into a 1D array.
-    # Includes non-iterable key-value pairs in the result.
-    def recursive_array_extractor(array_of_hashes)
-      rows = []
-      array_of_hashes.each do |hash_element|
-        rows << hash_element
-        hash_element.values.each do |values|
-          if eachable? values
-            child_rows = recursive_array_extractor(values)
-            rows.push(*child_rows)
+    # Generate a hash of settings for a single cell
+    def generate_smrt_cell_settings(well) # rubocop:disable Metrics/MethodLength
+      {
+        'Well Name'	=> well.used_aliquots.first.source.tube.barcode, # TRAC-2-7242
+        'Library Type'	=> well.library_type, # Standard
+        'Movie Acquisition Time (hours)'	=> well.movie_acquisition_time, # 24
+        'Insert Size (bp)'	=> well.insert_size, # 500
+        'Assign Data To Project'	=> 1, # (maybe we need to assign a run a project in traction)?
+        'Library Concentration (pM)'	=> well.library_concentration, # 250
+        'Include Base Kinetics'	=> well.include_base_kinetics,
+        'Polymerase Kit'	=> well.polymerase_kit, # 032037102739100071224
+        'Indexes'	=> well.barcode_set, # 244d96c6-f3b2-4997-5ae3-23ed33ab925f
+        'Sample is indexed'	=> well.tagged?, # Set to True to Multiplex
+        'Bio Sample Name' => well.tagged? ? nil : well.bio_sample_name,
+        'Use Adaptive Loading'	=> false, # this will likely be a well default in future
+        'Consensus Mode'	=> 'molecule', # (default to molecule do we need a custom field)
+        'Same Barcodes on Both Ends of Sequence'	=> well.same_barcodes_on_both_ends_of_sequence
+      }
+    end
+
+    # Generate a hash of settings for the cell
+    def generate_wells
+      object.plates.flat_map(&:wells).each_with_object({}) do |well, hash|
+        plate_well_name = "#{well.plate.plate_number}_#{well.position_leading_zero}"
+        hash[plate_well_name] = well
+      end
+    end
+
+    def generate_smrt_cells(wells)
+      wells.each_with_object({}) do |(plate_well, well), acc|
+        acc[plate_well] = generate_smrt_cell_settings(well)
+      end
+    end
+
+    def transpose_smrt_cells(smrt_cells)
+      smrt_cells.each_with_object({}) do |(plate_well, cell_data), result|
+        cell_data.each do |(key, value)|
+          result[key] ||= {}
+          result[key].merge!({ plate_well => value })
+        end
+      end
+    end
+
+    # Each key is a plate-well identifier and the value is a hash of settings for a particular cell
+    def smrt_cell_settings
+      wells = generate_wells
+      smrt_cells = generate_smrt_cells(wells)
+      transpose_smrt_cells(smrt_cells)
+    end
+
+    def sample_data(well, sample)
+      [
+        sample.bio_sample_name,
+        well.plate_well_position,
+        sample.adapter, # left adapter
+        sample.adapter  # right adapter
+      ]
+    end
+
+    # Generate a CSV of samples
+    def samples_csv
+      headers = ['Bio Sample Name', 'Plate Well', 'Adapter', 'Adapter2']
+      CSV.generate do |csv|
+        csv << headers
+        object.sorted_wells.each do |well|
+          well.aliquots_to_show_per_row&.each do |sample|
+            csv << sample_data(well, sample)
           end
         end
       end
-      rows
     end
 
-    # Parse the JSON data structure to derive the CSV. If a
-    # value contains an array, the key name will be ignored
-    # and the individual array elements parsed.
     def payload
-      CSV.generate do |csv|
-        csv << csv_headers
-        generate_csv_rows(csv)
-      end
-    end
+      # Combine all elements into the final sample sheet
+      sample_sheet = "[Run Settings]\n"
 
-    def generate_csv_rows(csv)
-      data_rows = recursive_array_extractor([data_structure])
-      # Initialize a flag to track if this is the first row
-      first_row = true
+      # Start with the run settings
+      sample_sheet += run_settings.map { |k, v| "#{k},#{v}" }.join("\n")
 
-      data_rows.each do |row|
-        row_values = process_row(row, first_row)
-        first_row = false if first_row && !row_values.all?(nil)
-        csv << row_values unless row_values.all?(nil)
-      end
-    end
+      # Add the cell settings
+      sample_sheet += "\n[SMRT Cell Settings]"
+      sample_sheet += ",#{plate_well_names.join(',')}\n"
+      sample_sheet += smrt_cell_settings.map do |key, cells|
+        "#{key}," + cells.values.join(',')
+      end.join("\n")
 
-    def process_row(row, first_row)
-      row_values = row.values_at(*csv_headers)
-      handle_csv_version(row_values, first_row)
-      row_values
-    end
+      # Add the sample settings
+      sample_sheet += "\n[Samples]\n"
+      sample_sheet += samples_csv
 
-    def handle_csv_version(row_values, first_row)
-      # CSV VERSION should be set only on first row
-      return if first_row
-
-      index = csv_headers.index('CSV Version')
-      row_values[index] = nil if index
+      sample_sheet
     end
   end
 end
