+++ config/pipelines/pacbio.yml
@@ -76,7 +76,7 @@ default: &default
     # For instructions on structure and syntax, see: 'app/exchanges/README.md'
     sample_sheet:
       v12_revio: &v12_revio
-        sample_sheet_class: PacbioSampleSheet
+        sample_sheet_class: PacbioSampleSheetLegacy
         column_order:
           - Library Type
           - Reagent Plate
@@ -172,7 +172,7 @@ default: &default
                   type: :model
                   value: bio_sample_name
       v12_sequel_iie: &v12_sequel_iie
-        sample_sheet_class: PacbioSampleSheet
+        sample_sheet_class: PacbioSampleSheetLegacy
         column_order:
           - Instrument Type
           - Run Name
@@ -294,7 +294,7 @@ default: &default
                   value: bio_sample_name
 
       v13_revio: &v13_revio
-        sample_sheet_class: PacbioSampleSheetV1
+        sample_sheet_class: PacbioSampleSheet
 
       # No changes required from v12_sequel_iie, added for consistency
       v13_sequel_iie: *v12_sequel_iie
