+++ spec/exchanges/run_csv/pacbio_sample_sheet_spec.rb
@@ -2,18 +2,102 @@
 
 require 'rails_helper'
 
+# import parser to assist with testing
+# huh? what is this? If we need a parser to parse the sample sheet probably needs a refactor
+require_relative '../../support/parsers/pacbio_sample_sheet_parser'
+
 RSpec.describe RunCsv::PacbioSampleSheet, type: :model do
+  before do
+    # Create a default smrt link version
+    create(:pacbio_smrt_link_version, name: 'v13_revio', default: true)
+  end
+
   describe '#payload' do
-    subject(:csv_string) { csv.payload }
+    subject(:sample_sheet_string) { sample_sheet.payload }
 
-    let(:plate)       { build(:pacbio_plate, wells:, plate_number: 1) }
-    let(:run)         { create(:pacbio_generic_run, smrt_link_version:, plates: [plate]) }
-    let(:parsed_csv)  { CSV.parse(csv_string) }
-    let(:csv)         { described_class.new(object: run, configuration:) }
+    let(:run)           { create(:pacbio_revio_run) }
+    let(:sample_sheet)  { described_class.new(object: run, configuration:) }
     let(:configuration) { Pipelines.pacbio.sample_sheet.by_version(run.smrt_link_version.name) }
+    let(:parsed_sample_sheet) { Parsers::PacbioSampleSheetParser.new.parse(sample_sheet_string) }
 
-    context 'v12_revio' do
-      let(:smrt_link_version) { create(:pacbio_smrt_link_version_default, name: 'v12_revio') }
+    context 'v13_revio' do
+      it 'must return a string' do
+        expect(sample_sheet_string.class).to eq String
+      end
+
+      it 'must have the three required sections' do
+        expect(sample_sheet_string).to include('[Run Settings]')
+        expect(sample_sheet_string).to include('[SMRT Cell Settings]')
+        expect(sample_sheet_string).to include('[Samples]')
+      end
+
+      it 'must have the three required sections in the correct order' do
+        run_settings_index = sample_sheet_string.index('[Run Settings]')
+        cell_settings_index = sample_sheet_string.index('[SMRT Cell Settings]')
+        samples_index = sample_sheet_string.index('[Samples]')
+
+        expect(run_settings_index).to be < cell_settings_index
+        expect(cell_settings_index).to be < samples_index
+      end
+
+      it 'must have the correct run settings' do
+        expect(parsed_sample_sheet['Run Settings']).to eq(
+          {
+            'Instrument Type' => 'Revio',
+            'Run Name' => run.name,
+            'Run Comments' => run.comments,
+            'Plate 1' => run.plates[0].sequencing_kit_box_barcode,
+            'Plate 2' => run.plates[1].sequencing_kit_box_barcode,
+            'CSV Version' => '1'
+          }
+        )
+      end
+
+      it 'must have the cells used listed on the same line as the section header' do
+        # get the line from sample_sheet_string containing [SMRT Cell Settings]
+        smrt_cell_settings_line = sample_sheet_string.lines.find { |line| line.include?('[SMRT Cell Settings]') }.strip
+        expect(smrt_cell_settings_line).to eq('[SMRT Cell Settings],1_A01,1_B01,2_A01')
+      end
+
+      it 'must have the correct SMRT cell settings' do
+        smrt_cell_settings = parsed_sample_sheet['SMRT Cell Settings']
+
+        # create a hash of plate_well_name => well for easy comparison
+        plate_wells = run.plates.flat_map(&:wells).each_with_object({}) do |well, hash|
+          plate_well_name = "#{well.plate.plate_number}_#{well.position_leading_zero}"
+          hash[plate_well_name] = well
+        end
+
+        # confirm that the wells are as expected
+        plate_well_names = plate_wells.keys
+        expect(plate_well_names).to contain_exactly('1_A01', '1_B01', '2_A01')
+        expect(smrt_cell_settings.keys).to match_array(plate_well_names)
+
+        plate_well_names.each do |plate_well_name|
+          well = plate_wells[plate_well_name]
+          expected_settings = {
+            # for all wells
+            'Well Name' => well.used_aliquots.first.source.tube.barcode,
+            'Library Type' => 'Standard',
+            'Movie Acquisition Time (hours)' => well.movie_acquisition_time.to_s,
+            'Insert Size (bp)' => well.insert_size.to_s,
+            'Assign Data To Project' => '1',
+            'Library Concentration (pM)' => well.library_concentration.to_s,
+            'Include Base Kinetics' => well.include_base_kinetics.downcase == 'true', # is a string
+            'Polymerase Kit' => well.polymerase_kit,
+            'Use Adaptive Loading' => false,
+            'Consensus Mode' => 'molecule',
+
+            # specific to tagged wells
+            'Bio Sample Name' => '',
+            'Sample is indexed' => true,
+            'Indexes' => well.barcode_set,
+            'Same Barcodes on Both Ends of Sequence' => true
+          }
+
+          expect(smrt_cell_settings[plate_well_name]).to eq(expected_settings)
+        end
+      end
 
       context 'when the libraries are tagged' do
         let(:well1) do
@@ -21,7 +105,9 @@ RSpec.describe RunCsv::PacbioSampleSheet, type: :model do
             :pacbio_well,
             pre_extension_time: 2,
             generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'Yes'
+            ccs_analysis_output: 'Yes',
+            row: 'A',
+            column: 1
           )
         end
         let(:well2) do
@@ -29,85 +115,58 @@ RSpec.describe RunCsv::PacbioSampleSheet, type: :model do
             :pacbio_well,
             pre_extension_time: 2,
             generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'No'
+            ccs_analysis_output: 'No',
+            row: 'A',
+            column: 1
           )
         end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string.class).to eq String
+        let(:well3) do
+          create(
+            :pacbio_well,
+            pre_extension_time: 2,
+            generate_hifi: 'In SMRT Link',
+            ccs_analysis_output: 'No',
+            row: 'B',
+            column: 1
+          )
         end
+        let(:plate1_wells)   { [well1] }
+        let(:plate2_wells)   { [well2, well3] }
+        let(:plate1)  { build(:pacbio_plate, wells: plate1_wells, plate_number: 1) }
+        let(:plate2)  { build(:pacbio_plate, wells: plate2_wells, plate_number: 2) }
+        let(:run)     { create(:pacbio_revio_run, plates: [plate1, plate2]) }
 
         it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
+          # get the line from sample_sheet_string after the one containing [Samples]
+          sample_sheet_lines = sample_sheet_string.lines
+          samples_section_index = sample_sheet_lines.find_index { |line| line.include?('[Samples]') }
+          headers_line = sample_sheet_lines[samples_section_index + 1]
+          headers = headers_line.strip.split(',')
+          expected_headers = ['Bio Sample Name', 'Plate Well', 'Adapter', 'Adapter2']
           expect(headers).to eq(expected_headers)
         end
 
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          well_data_2 = parsed_csv[7]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well1],
-            [well_data_2, well2]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              'Standard', # library type
-              '1', # reagent plate
-              well.plate.sequencing_kit_box_barcode,
-              nil, # plate 2: sequencing_kit_box_barcode
-              well.plate.run.name,
-              well.plate.run.system_name,
-              well.plate.run.comments,
-              'true', # well.collection?
-              well.position_leading_zero,
-              well.tube_barcode,
-              well.movie_acquisition_time.to_s,
-              well.include_base_kinetics.to_s,
-              well.library_concentration.to_s,
-              well.polymerase_kit,
-              well.automation_parameters,
-              well.barcode_set,
-              nil, # barcode name - does not apply
-              nil # sample name - does not apply
-            ])
-          end
-        end
-
         it 'must have the correct sample rows' do
-          # Note the increment in the parsed_csv index
-          sample_data_1 = parsed_csv[2]
-          sample_data_2 = parsed_csv[8]
+          # 5 pools per well
+          sample_data_1 = parsed_sample_sheet['Samples'][0]
+          sample_data_2 = parsed_sample_sheet['Samples'][1 * 5]
+          sample_data_3 = parsed_sample_sheet['Samples'][2 * 5]
 
           #  iterate through the samples under test
           sample_expectations = [
             [sample_data_1, well1],
-            [sample_data_2, well2]
+            [sample_data_2, well2],
+            [sample_data_3, well3]
           ]
           sample_expectations.each do |sample_data, well|
-            expect(sample_data).to eq([
-              nil,
-              '1', # reagent plate
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              'false', # aliquot.collection?
-              well.position,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              well.base_used_aliquots.first.barcode_name,
-              well.base_used_aliquots.first.bio_sample_name
-            ])
+            expect(sample_data).to eq(
+              {
+                'Bio Sample Name' => well.base_used_aliquots.first.bio_sample_name,
+                'Plate Well' => well.plate_well_position,
+                'Adapter' => well.base_used_aliquots.first.tag.group_id,
+                'Adapter2' => well.base_used_aliquots.first.tag.group_id
+              }
+            )
           end
         end
       end
@@ -115,76 +174,16 @@ RSpec.describe RunCsv::PacbioSampleSheet, type: :model do
       context 'when the libraries are untagged' do
         let(:pool1)   { create_list(:pacbio_pool, 1, :untagged) }
         let(:pool2)   { create_list(:pacbio_pool, 1, :untagged) }
-        let(:well1)   do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: pool1)
-        end
-        let(:well2) do
-          create(:pacbio_well, generate_hifi: 'On Instrument', ccs_analysis_output: 'No',
-                               pools: pool2)
-        end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          well_data_2 = parsed_csv[2]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well1],
-            [well_data_2, well2]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              'Standard', # library type
-              '1', # reagent plate
-              well.plate.sequencing_kit_box_barcode,
-              nil, # plate 2: sequencing_kit_box_barcode
-              well.plate.run.name,
-              well.plate.run.system_name,
-              well.plate.run.comments,
-              'true', # well.collection?
-              well.position_leading_zero,
-              well.tube_barcode,
-              well.movie_acquisition_time.to_s,
-              well.include_base_kinetics.to_s,
-              well.library_concentration.to_s,
-              well.polymerase_kit,
-              well.automation_parameters,
-              well.barcode_set,
-              nil, # barcode name - does not apply
-              well.bio_sample_name
-            ])
-          end
-        end
-
-        it 'must not have sample rows' do
-          expect(parsed_csv.size).to eq 3
-        end
-      end
-    end
-
-    context 'v12_sequel_iie and v13_sequel_iie' do
-      let(:smrt_link_version) { create(:pacbio_smrt_link_version, name: 'v12_sequel_iie', default: true) }
-      let(:run)               { create(:pacbio_sequel_run, smrt_link_version:, plates: [plate]) }
-
-      context 'when the libraries are tagged' do
+        let(:pool3)   { create_list(:pacbio_pool, 1, :untagged) }
         let(:well1) do
           create(
             :pacbio_well,
             pre_extension_time: 2,
             generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'Yes'
+            ccs_analysis_output: 'Yes',
+            pools: pool1, # untagged pool
+            row: 'A',
+            column: 1
           )
         end
         let(:well2) do
@@ -192,734 +191,82 @@ RSpec.describe RunCsv::PacbioSampleSheet, type: :model do
             :pacbio_well,
             pre_extension_time: 2,
             generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'No'
+            ccs_analysis_output: 'No',
+            pools: pool2, # untagged pool
+            row: 'A',
+            column: 1
           )
         end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string.class).to eq String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          well_data_2 = parsed_csv[7]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well1],
-            [well_data_2, well2]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true', # well.collection?
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.show_row_per_sample?.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              nil, # sample name - does not apply
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must have the correct sample rows' do
-          # Note the increment in the parsed_csv index
-          sample_data_1 = parsed_csv[2]
-          sample_data_2 = parsed_csv[8]
-
-          #  iterate through the samples under test
-          sample_expectations = [
-            [sample_data_1, well1],
-            [sample_data_2, well2]
-          ]
-          sample_expectations.each do |sample_data, well|
-            expect(sample_data).to eq([
-              nil,
-              nil,
-              'false', # aliquot.collection?
-              well.position,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              well.base_used_aliquots.first.barcode_name,
-              nil,
-              nil,
-              well.base_used_aliquots.first.bio_sample_name,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil
-            ])
-          end
-        end
-      end
-
-      context 'when the libraries are untagged' do
-        let(:pool1)   { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool2)   { create_list(:pacbio_pool, 1, :untagged) }
-        let(:well1)   do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: pool1)
-        end
-        let(:well2) do
-          create(:pacbio_well, generate_hifi: 'On Instrument', ccs_analysis_output: 'No',
-                               pools: pool2)
-        end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          well_data_2 = parsed_csv[2]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well1],
-            [well_data_2, well2]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true', # well.collection?
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must not have sample rows' do
-          expect(parsed_csv.size).to eq 3
-        end
-      end
-
-      context 'when the run has libraries and pools' do
-        let(:pool) { create(:pacbio_pool, :tagged, library_count: 1) }
-        let(:library) { create(:pacbio_library, :tagged) }
-        let(:well1)   do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: [pool])
-        end
-        let(:well2) do
-          create(:pacbio_well, generate_hifi: 'On Instrument', ccs_analysis_output: 'No',
-                               libraries: [library], pool_count: 0)
-        end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          well_data_2 = parsed_csv[3]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well1],
-            [well_data_2, well2]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true', # well.collection?
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must not have sample rows' do
-          expect(parsed_csv.size).to eq 5
-        end
-      end
-
-      context 'when the run has pools that use libraries and requests' do
-        let(:pool) { create(:pacbio_pool, :tagged, library_count: 1) }
-        let(:well) do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: [pool])
-        end
-        let(:wells) { [well] }
-
-        before do
-          pool.used_aliquots << create(:aliquot, source: create(:pacbio_library))
-        end
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_data_1 = parsed_csv[1]
-          #  iterate through the wells under test
-          well_expectations = [
-            [well_data_1, well]
-          ]
-          well_expectations.each do |well_data, well|
-            expect(well_data).to eq([
-              well.plate.run.system_name,
-              well.plate.run.name,
-              well.collection?.to_s,
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must have the correct sample rows' do
-          # Note the increment in the parsed_csv index
-          sample_data_1 = parsed_csv[2]
-          sample_data_2 = parsed_csv[3]
-
-          # iterate through the samples under test
-          sample_expectations = [
-            [sample_data_1, well, 0],
-            [sample_data_2, well, 1]
-          ]
-          sample_expectations.each do |sample_data, well, aliquot_index|
-            expect(sample_data).to eq([
-              nil,
-              nil,
-              'false', # aliquot.collection?
-              well.position,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              well.base_used_aliquots[aliquot_index].barcode_name,
-              nil,
-              nil,
-              well.base_used_aliquots[aliquot_index].bio_sample_name,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil
-            ])
-          end
-        end
-      end
-
-      context 'with lots of wells in unpredictable orders' do
-        let(:pool1) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool2) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool3) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:well1) { create(:pacbio_well, pools: pool1, row: 'A', column: 10) }
-        let(:well2) { create(:pacbio_well, pools: pool2, row: 'A', column: 5) }
-        let(:well3) { create(:pacbio_well, pools: pool3, row: 'B', column: 1) }
-        let(:wells) { [well1, well2, well3] }
-
-        it 'sorts the wells by column' do
-          sorted_well_positions = parsed_csv[1..].pluck(3)
-          expect(sorted_well_positions).to eq(%w[B01 A05 A10])
-        end
-      end
-    end
-
-    context 'v13_1_sequel_iie' do
-      let(:smrt_link_version) { create(:pacbio_smrt_link_version, name: 'v13_1_sequel_iie', default: true) }
-      let(:run)               { create(:pacbio_sequel_run, smrt_link_version:, plates: [plate]) }
-
-      context 'when the libraries are tagged' do
-        let(:well1) do
+        let(:well3) do
           create(
             :pacbio_well,
             pre_extension_time: 2,
             generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'Yes'
+            ccs_analysis_output: 'No',
+            pools: pool3, # untagged pool
+            row: 'B',
+            column: 1
           )
         end
-        let(:well2) do
-          create(
-            :pacbio_well,
-            pre_extension_time: 2,
-            generate_hifi: 'In SMRT Link',
-            ccs_analysis_output: 'No'
-          )
-        end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string.class).to eq String
-        end
+        let(:plate1_wells)   { [well1] }
+        let(:plate2_wells)   { [well2, well3] }
+        let(:plate1)  { build(:pacbio_plate, wells: plate1_wells, plate_number: 1) }
+        let(:plate2)  { build(:pacbio_plate, wells: plate2_wells, plate_number: 2) }
+        let(:run)     { create(:pacbio_revio_run, plates: [plate1, plate2]) }
 
         it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
+          # get the line from sample_sheet_string after the one containing [Samples]
+          sample_sheet_lines = sample_sheet_string.lines
+          samples_section_index = sample_sheet_lines.find_index { |line| line.include?('[Samples]') }
+          headers_line = sample_sheet_lines[samples_section_index + 1]
+          headers = headers_line.strip.split(',')
+          expected_headers = ['Bio Sample Name', 'Plate Well', 'Adapter', 'Adapter2']
           expect(headers).to eq(expected_headers)
         end
 
-        it 'must have the correct well header rows' do
-          # Define expectations for each well's data
-          well_expectations = [
-            { data: parsed_csv[1], well: well1, csv_version: '1' },
-            { data: parsed_csv[7], well: well2, csv_version: nil }
-          ]
+        it 'must have the wells added to the SMRT Cell Settings section' do
+          smrt_cell_settings = parsed_sample_sheet['SMRT Cell Settings']
 
-          # Iterate through each well expectation and compare
-          well_expectations.each do |expectation|
-            well_data = expectation[:data]
-            well = expectation[:well]
-            csv_version = expectation[:csv_version]
-
-            expected_values = [
-              csv_version,
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true', # well.collection?
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.show_row_per_sample?.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              nil, # sample name - does not apply
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ]
-
-            expect(well_data).to eq(expected_values)
+          # create a hash of plate_well_name => well for easy comparison
+          plate_wells = run.plates.flat_map(&:wells).each_with_object({}) do |well, hash|
+            plate_well_name = "#{well.plate.plate_number}_#{well.position_leading_zero}"
+            hash[plate_well_name] = well
           end
-        end
 
-        it 'must have the correct sample rows' do
-          # Note the increment in the parsed_csv index
-          sample_data_1 = parsed_csv[2]
-          sample_data_2 = parsed_csv[8]
+          # confirm that the wells are as expected
+          plate_well_names = plate_wells.keys
+          expect(plate_well_names).to contain_exactly('1_A01', '2_A01', '2_B01')
+          expect(smrt_cell_settings.keys).to match_array(plate_well_names)
 
-          #  iterate through the samples under test
-          sample_expectations = [
-            [sample_data_1, well1],
-            [sample_data_2, well2]
-          ]
-          sample_expectations.each do |sample_data, well|
-            expect(sample_data).to eq([
-              nil,
-              nil,
-              nil,
-              'false', # aliquot.collection?
-              well.position,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              well.base_used_aliquots.first.barcode_name,
-              nil,
-              nil,
-              well.base_used_aliquots.first.bio_sample_name,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil
-            ])
-          end
-        end
-      end
+          #  iterate through the wells under test
+          plate_well_names.each do |plate_well_name|
+            well_data = smrt_cell_settings[plate_well_name]
+            well = plate_wells[plate_well_name]
 
-      context 'when the libraries are untagged' do
-        let(:pool1)   { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool2)   { create_list(:pacbio_pool, 1, :untagged) }
-        let(:well1)   do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: pool1)
-        end
-        let(:well2) do
-          create(:pacbio_well, generate_hifi: 'On Instrument', ccs_analysis_output: 'No',
-                               pools: pool2)
-        end
-        let(:wells) { [well1, well2] }
+            expect(well_data).to eq(
+              # for all wells
+              'Well Name' => well.used_aliquots.first.source.tube.barcode,
+              'Library Type' => 'Standard',
+              'Movie Acquisition Time (hours)' => well.movie_acquisition_time.to_s,
+              'Insert Size (bp)' => well.insert_size.to_s,
+              'Assign Data To Project' => '1',
+              'Library Concentration (pM)' => well.library_concentration.to_s,
+              'Include Base Kinetics' => well.include_base_kinetics.downcase == 'true', # is a string
+              'Polymerase Kit' => well.polymerase_kit,
+              'Use Adaptive Loading' => false,
+              'Consensus Mode' => 'molecule',
 
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_expectations = [
-            { data: parsed_csv[1], well: well1, csv_version: '1' },
-            { data: parsed_csv[2], well: well2, csv_version: nil }
-          ]
-          well_expectations.each do |expectation|
-            well_data = expectation[:data]
-            well = expectation[:well]
-            csv_version = expectation[:csv_version]
-            expect(well_data).to eq([
-              csv_version,
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true',
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
+              # specific to untagged wells
+              'Bio Sample Name' => well.bio_sample_name,
+              'Sample is indexed' => false,
+              'Indexes' => '', # well.barcode_set
+              'Same Barcodes on Both Ends of Sequence' => '' # well.same_barcodes_on_both_ends_of_sequence.to_s
+            )
           end
         end
 
         it 'must not have sample rows' do
-          expect(parsed_csv.size).to eq 3
-        end
-      end
-
-      context 'when the run has libraries and pools' do
-        let(:pool) { create(:pacbio_pool, :tagged, library_count: 1) }
-        let(:library) { create(:pacbio_library, :tagged) }
-        let(:well1)   do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: [pool])
-        end
-        let(:well2) do
-          create(:pacbio_well, generate_hifi: 'On Instrument', ccs_analysis_output: 'No',
-                               libraries: [library], pool_count: 0)
-        end
-        let(:wells) { [well1, well2] }
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_expectations = [
-            { data: parsed_csv[1], well: well1, csv_version: '1' },
-            { data: parsed_csv[3], well: well2, csv_version: nil }
-          ]
-
-          well_expectations.each do |expectation|
-            well_data = expectation[:data]
-            well = expectation[:well]
-            csv_version = expectation[:csv_version]
-            expect(well_data).to eq([
-              csv_version,
-              well.plate.run.system_name,
-              well.plate.run.name,
-              'true',
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must not have sample rows' do
-          expect(parsed_csv.size).to eq 5
-        end
-      end
-
-      context 'when the run has pools that use libraries and requests' do
-        let(:pool) { create(:pacbio_pool, :tagged, library_count: 1) }
-        let(:well) do
-          create(:pacbio_well, pre_extension_time: 2, generate_hifi: 'Do Not Generate',
-                               ccs_analysis_output: 'Yes', pools: [pool])
-        end
-        let(:wells) { [well] }
-
-        before do
-          pool.used_aliquots << create(:aliquot, source: create(:pacbio_library))
-        end
-
-        it 'must return a csv string' do
-          expect(csv_string).to be_a String
-        end
-
-        it 'must have the correct headers' do
-          headers = parsed_csv[0]
-
-          expected_headers = configuration.column_order
-          expect(headers).to eq(expected_headers)
-        end
-
-        it 'must have the correct well header rows' do
-          well_expectations = [
-            { data: parsed_csv[1], well:, csv_version: '1' }
-          ]
-          well_expectations.each do |expectation|
-            well_data = expectation[:data]
-            well = expectation[:well]
-            csv_version = expectation[:csv_version]
-            expect(well_data).to eq([
-              csv_version,
-              well.plate.run.system_name,
-              well.plate.run.name,
-              well.collection?.to_s,
-              well.position,
-              well.tube_barcode,
-              well.movie_time.to_s,
-              well.insert_size.to_s,
-              well.template_prep_kit_box_barcode,
-              well.binding_kit_box_barcode,
-              well.plate.sequencing_kit_box_barcode,
-              well.on_plate_loading_concentration.to_s,
-              well.plate.run.dna_control_complex_box_barcode,
-              well.plate.run.comments,
-              well.sample_is_barcoded.to_s,
-              nil, # barcode name - does not apply
-              well.barcode_set,
-              well.same_barcodes_on_both_ends_of_sequence.to_s,
-              well.bio_sample_name,
-              well.automation_parameters,
-              well.ccs_analysis_output_include_kinetics_information,
-              well.loading_target_p1_plus_p2.to_s,
-              well.adaptive_loading_check.to_s,
-              well.ccs_analysis_output_include_low_quality_reads,
-              well.include_fivemc_calls_in_cpg_motifs,
-              well.demultiplex_barcodes
-            ])
-          end
-        end
-
-        it 'must have the correct sample rows' do
-          sample_data_1 = parsed_csv[2]
-          sample_data_2 = parsed_csv[3]
-
-          sample_expectations = [
-            [sample_data_1, well, 0],
-            [sample_data_2, well, 1]
-          ]
-          sample_expectations.each do |sample_data, well, aliquot_index|
-            expect(sample_data).to eq([
-              nil,
-              nil,
-              nil,
-              'false', # aliquot.collection?
-              well.position,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              well.base_used_aliquots[aliquot_index].barcode_name,
-              nil,
-              nil,
-              well.base_used_aliquots[aliquot_index].bio_sample_name,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil,
-              nil
-            ])
-          end
-        end
-      end
-
-      context 'with lots of wells in unpredictable orders' do
-        let(:pool1) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool2) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:pool3) { create_list(:pacbio_pool, 1, :untagged) }
-        let(:well1) { create(:pacbio_well, pools: pool1, row: 'A', column: 10) }
-        let(:well2) { create(:pacbio_well, pools: pool2, row: 'A', column: 5) }
-        let(:well3) { create(:pacbio_well, pools: pool3, row: 'B', column: 1) }
-        let(:wells) { [well1, well2, well3] }
-
-        it 'sorts the wells by column' do
-          sorted_well_positions = parsed_csv[1..].pluck(4)
-          expect(sorted_well_positions).to eq(%w[B01 A05 A10])
+          expect(parsed_sample_sheet['Samples']).to be_empty
         end
       end
     end
